+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-### vNEXT
+### v0.9.16
+- Add ability to set custom WebSocket protocols for client. <br/>
+  [@pkosiec](https://github.com/pkosiec) in [#477](https://github.com/apollographql/subscriptions-transport-ws/pull/477)
+
+### v0.9.15
+
+- Add support for `graphql` and `@types/graphql` 14.  <br/>
+  [@caiquecastro](https://github.com/caiquecastro) in [#464](https://github.com/apollographql/subscriptions-transport-ws/pull/464)
 
 ### v0.9.13
 - Allow connectionParams to be a Promise [PR #443](https://github.com/apollographql/subscriptions-transport-ws/pull/443)